Port prometheus to docker
